+++ picasso/gui/render.py
@@ -13,7 +13,6 @@ import copy
 import time
 import os.path
 import importlib, pkgutil
-from glob import glob
 from math import ceil
 from functools import partial
 
@@ -208,10 +207,7 @@ def check_picks(f):
             QtWidgets.QMessageBox.information(
                 args[0],
                 "Pick Error",
-                (
-                    "No localizations picked."
-                    " Please pick at least twice first."
-                ),
+                "Please pick at least twice.",
             )
         else:
             return f(args[0])
@@ -1417,11 +1413,10 @@ class ClsDlg3D(QtWidgets.QDialog):
             + "."
         )
 
-        print("Mode 1")
         locs = all_picked_locs[current]
         locs = stack_arrays(locs, asrecarray=True, usemask=False)
 
-        est = KMeans(n_clusters=n_clusters)
+        est = KMeans(n_clusters=n_clusters, n_init='auto')
 
         scaled_locs = lib.append_to_rec(
             locs, locs["x"] * pixelsize, "x_scaled"
@@ -1441,7 +1436,9 @@ class ClsDlg3D(QtWidgets.QDialog):
         counts = list(Counter(labels).items())
         # l_locs = lib.append_to_rec(l_locs,labels,'cluster')
 
-        ax1.scatter(locs["x"], locs["y"], locs["z"], c=labels.astype(np.float), s=2)
+        ax1.scatter(
+            locs["x"], locs["y"], locs["z"], c=labels.astype(float), s=2
+        )
 
         ax1.set_xlabel("X")
         ax1.set_ylabel("Y")
@@ -1615,11 +1612,10 @@ class ClsDlg2D(QtWidgets.QDialog):
             + "."
         )
 
-        print("Mode 1")
         locs = all_picked_locs[current]
         locs = stack_arrays(locs, asrecarray=True, usemask=False)
 
-        est = KMeans(n_clusters=n_clusters)
+        est = KMeans(n_clusters=n_clusters, n_init='auto')
 
         scaled_locs = lib.append_to_rec(locs, locs["x"], "x_scaled")
         scaled_locs = lib.append_to_rec(scaled_locs, locs["y"], "y_scaled")
@@ -1634,7 +1630,7 @@ class ClsDlg2D(QtWidgets.QDialog):
         counts = list(Counter(labels).items())
         # l_locs = lib.append_to_rec(l_locs,labels,'cluster')
 
-        ax1.scatter(locs["x"], locs["y"], c=labels.astype(np.float), s=2)
+        ax1.scatter(locs["x"], locs["y"], c=labels.astype(float), s=2)
 
         ax1.set_xlabel("X")
         ax1.set_ylabel("Y")
@@ -1790,7 +1786,7 @@ class DbscanDialog(QtWidgets.QDialog):
         grid.addWidget(self.radius, 0, 1)
         grid.addWidget(QtWidgets.QLabel("Min. samples:"), 1, 0)
         self.density = QtWidgets.QSpinBox()
-        self.density.setRange(1, 1e6)
+        self.density.setRange(1, int(1e6))
         self.density.setValue(4)
         grid.addWidget(self.density, 1, 1)
         vbox.addLayout(grid)
@@ -1859,12 +1855,12 @@ class HdbscanDialog(QtWidgets.QDialog):
         grid = QtWidgets.QGridLayout()
         grid.addWidget(QtWidgets.QLabel("Min. cluster size:"), 0, 0)
         self.min_cluster = QtWidgets.QSpinBox()
-        self.min_cluster.setRange(1, 1e6)
+        self.min_cluster.setRange(1, int(1e6))
         self.min_cluster.setValue(10)
         grid.addWidget(self.min_cluster, 0, 1)
         grid.addWidget(QtWidgets.QLabel("Min. samples:"), 1, 0)
         self.min_samples = QtWidgets.QSpinBox()
-        self.min_samples.setRange(1, 1e6)
+        self.min_samples.setRange(1, int(1e6))
         self.min_samples.setValue(10)
         grid.addWidget(self.min_samples, 1, 1)
         grid.addWidget(QtWidgets.QLabel(
@@ -1957,7 +1953,7 @@ class SMLMDialog3D(QtWidgets.QDialog):
         # min no. locs
         grid.addWidget(QtWidgets.QLabel("Min. no. locs:"), 2, 0)
         self.min_locs = QtWidgets.QSpinBox()
-        self.min_locs.setRange(1, 1e6)
+        self.min_locs.setRange(1, int(1e6))
         self.min_locs.setValue(10)
         grid.addWidget(self.min_locs, 2, 1)
         # save cluster centers
@@ -2039,7 +2035,7 @@ class SMLMDialog2D(QtWidgets.QDialog):
         # min no. locs
         grid.addWidget(QtWidgets.QLabel("Min. no. locs:"), 1, 0)
         self.min_locs = QtWidgets.QSpinBox()
-        self.min_locs.setRange(1, 1e6)
+        self.min_locs.setRange(1, int(1e6))
         self.min_locs.setValue(10)
         grid.addWidget(self.min_locs, 1, 1)
         # save cluster centers
@@ -2467,7 +2463,7 @@ class TestDBSCANParams(QtWidgets.QWidget):
         self.min_samples = QtWidgets.QSpinBox()
         self.min_samples.setKeyboardTracking(False)
         self.min_samples.setValue(4)
-        self.min_samples.setRange(1, 1e6)
+        self.min_samples.setRange(1, int(1e6))
         self.min_samples.setSingleStep(1)
         grid.addWidget(self.min_samples, 1, 1)
         grid.setRowStretch(2, 1)
@@ -2486,7 +2482,7 @@ class TestHDBSCANParams(QtWidgets.QWidget):
         self.min_cluster_size = QtWidgets.QSpinBox()
         self.min_cluster_size.setKeyboardTracking(False)
         self.min_cluster_size.setValue(10)
-        self.min_cluster_size.setRange(1, 1e6)
+        self.min_cluster_size.setRange(1, int(1e6))
         self.min_cluster_size.setSingleStep(1)
         grid.addWidget(self.min_cluster_size, 0, 1)
 
@@ -2494,7 +2490,7 @@ class TestHDBSCANParams(QtWidgets.QWidget):
         self.min_samples = QtWidgets.QSpinBox()
         self.min_samples.setKeyboardTracking(False)
         self.min_samples.setValue(10)
-        self.min_samples.setRange(1, 1e6)
+        self.min_samples.setRange(1, int(1e6))
         self.min_samples.setSingleStep(1)
         grid.addWidget(self.min_samples, 1, 1)
 
@@ -2540,7 +2536,7 @@ class TestSMLMParams(QtWidgets.QWidget):
         self.min_locs = QtWidgets.QSpinBox()
         self.min_locs.setKeyboardTracking(False)
         self.min_locs.setValue(10)
-        self.min_locs.setRange(1, 1e6)
+        self.min_locs.setRange(1, int(1e6))
         self.min_locs.setSingleStep(1)
         grid.addWidget(self.min_locs, 2, 1)
 
@@ -3264,7 +3260,7 @@ class InfoDialog(QtWidgets.QDialog):
             QtWidgets.QLabel("Ignore dark times <="), row, 0
         )
         self.max_dark_time = QtWidgets.QSpinBox()
-        self.max_dark_time.setRange(0, 1e9)
+        self.max_dark_time.setRange(0, int(1e9))
         self.max_dark_time.setValue(1)
         self.picks_grid.addWidget(self.max_dark_time, row, 1, 1, 2)
         row = self.picks_grid.rowCount()
@@ -3284,7 +3280,7 @@ class InfoDialog(QtWidgets.QDialog):
             QtWidgets.QLabel("# Units per pick:"), row, 0
         )
         self.units_per_pick = QtWidgets.QSpinBox()
-        self.units_per_pick.setRange(1, 1e6)
+        self.units_per_pick.setRange(1, int(1e6))
         self.units_per_pick.setValue(1)
         self.picks_grid.addWidget(self.units_per_pick, row, 1, 1, 2)
         calculate_influx_button = QtWidgets.QPushButton("Calibrate influx")
@@ -3537,7 +3533,7 @@ class MaskSettingsDialog(QtWidgets.QDialog):
 
         mask_grid.addWidget(QtWidgets.QLabel("Display pixel size [nm]"), 0, 0)
         self.disp_px_size = QtWidgets.QSpinBox()
-        self.disp_px_size.setRange(0.1, 99999)
+        self.disp_px_size.setRange(10, 99999)
         self.disp_px_size.setValue(300)
         self.disp_px_size.setSingleStep(10)
         self.disp_px_size.setKeyboardTracking(False)
@@ -3903,7 +3899,7 @@ class PickToolCircleSettings(QtWidgets.QWidget):
         self.window = window
         self.grid.addWidget(QtWidgets.QLabel("Diameter (cam. pixel):"), 0, 0)
         self.pick_diameter = QtWidgets.QDoubleSpinBox()
-        self.pick_diameter.setRange(0, 999999)
+        self.pick_diameter.setRange(0.001, 999999)
         self.pick_diameter.setValue(1)
         self.pick_diameter.setSingleStep(0.1)
         self.pick_diameter.setDecimals(3)
@@ -3932,7 +3928,7 @@ class PickToolRectangleSettings(QtWidgets.QWidget):
         self.grid = QtWidgets.QGridLayout(self)
         self.grid.addWidget(QtWidgets.QLabel("Width (cam. pixel):"), 0, 0)
         self.pick_width = QtWidgets.QDoubleSpinBox()
-        self.pick_width.setRange(0, 999999)
+        self.pick_width.setRange(0.001, 999999)
         self.pick_width.setValue(1)
         self.pick_width.setSingleStep(0.1)
         self.pick_width.setDecimals(3)
@@ -4071,7 +4067,7 @@ class RESIDialog(QtWidgets.QDialog):
             self.radius_z.append(r_z)
 
             min_locs = QtWidgets.QSpinBox()
-            min_locs.setRange(1, 1e6)
+            min_locs.setRange(1, int(1e6))
             min_locs.setValue(10)
             min_locs.setSingleStep(1)
             self.min_locs.append(min_locs)
@@ -5061,8 +5057,8 @@ class SlicerDialog(QtWidgets.QDialog):
         self.ax.set_ylabel("Rel. frequency")
         self.ax.set_title(r"$\mathrm{Histogram\ of\ Z:}$")
         self.canvas.draw()
-        self.sl.setMaximum(len(self.bins) - 2)
-        self.sl.setValue(len(self.bins) / 2)
+        self.sl.setMaximum(int(len(self.bins)) - 2)
+        self.sl.setValue(int(len(self.bins) / 2))
 
         # reset cache
         self.slicer_cache = {}
@@ -5076,7 +5072,7 @@ class SlicerDialog(QtWidgets.QDialog):
             self.calculate_histogram()
         else:
             self.calculate_histogram()
-            self.sl.setValue(len(self.bins) / 2)
+            self.sl.setValue(int(len(self.bins) / 2))
             # self.on_slice_position_changed(self.sl.value())
 
     def toggle_slicer(self):
@@ -6438,14 +6434,14 @@ class View(QtWidgets.QLabel):
             alpha = np.arctan((end_y - start_y) / (end_x - start_x))
         dx = width * np.sin(alpha) / 2
         dy = width * np.cos(alpha) / 2
-        x1 = start_x - dx
-        x2 = start_x + dx
-        x4 = end_x - dx
-        x3 = end_x + dx
-        y1 = start_y + dy
-        y2 = start_y - dy
-        y4 = end_y + dy
-        y3 = end_y - dy
+        x1 = int(start_x - dx)
+        x2 = int(start_x + dx)
+        x4 = int(end_x - dx)
+        x3 = int(end_x + dx)
+        y1 = int(start_y + dy)
+        y2 = int(start_y - dy)
+        y4 = int(end_y + dy)
+        y3 = int(end_y - dy)
         return [x1, x2, x3, x4], [y1, y2, y3, y4]
 
     def get_pick_rectangle_polygon(
@@ -6520,6 +6516,7 @@ class View(QtWidgets.QLabel):
             else:
                 d = t_dialog.pick_diameter.value()
                 d *= self.width() / self.viewport_width()
+                d = int(d)
 
                 painter = QtGui.QPainter(image)
                 painter.setPen(QtGui.QColor("yellow"))
@@ -6532,11 +6529,13 @@ class View(QtWidgets.QLabel):
 
                     # convert from camera units to display units
                     cx, cy = self.map_to_view(*pick)
-                    painter.drawEllipse(cx - d / 2, cy - d / 2, d, d)
+                    painter.drawEllipse(int(cx - d / 2), int(cy - d / 2), d, d)
 
                     # annotate picks
                     if t_dialog.pick_annotation.isChecked():
-                        painter.drawText(cx + d / 2, cy + d / 2, str(i))
+                        painter.drawText(
+                            int(cx + d / 2), int(cy + d / 2), str(i)
+                        )
                 painter.end()
 
         # draw rectangular picks
@@ -6654,10 +6653,10 @@ class View(QtWidgets.QLabel):
 
             # draw a cross
             painter.drawPoint(cx, cy)
-            painter.drawLine(cx, cy, cx + d / 2, cy)
-            painter.drawLine(cx, cy, cx, cy + d / 2)
-            painter.drawLine(cx, cy, cx - d / 2, cy)
-            painter.drawLine(cx, cy, cx, cy - d / 2)
+            painter.drawLine(cx, cy, int(cx + d / 2), cy)
+            painter.drawLine(cx, cy, cx, int(cy + d / 2))
+            painter.drawLine(cx, cy, int(cx - d / 2), cy)
+            painter.drawLine(cx, cy, cx, int(cy - d / 2))
 
             # draw a line between points and show distance
             if oldpoint != []:
@@ -6683,7 +6682,9 @@ class View(QtWidgets.QLabel):
                     / 100
                 )
                 painter.drawText(
-                    (cx + ox) / 2 + d, (cy + oy) / 2 + d, str(distance) + " nm"
+                    int((cx + ox) / 2 + d), 
+                    int((cy + oy) / 2 + d), 
+                    str(distance) + " nm",
                 )
             oldpoint = point
         painter.end()
@@ -6806,7 +6807,7 @@ class View(QtWidgets.QLabel):
         if self.window.display_settings_dlg.minimap.isChecked():
             movie_height, movie_width = self.movie_size()
             length_minimap = 100
-            height_minimap = movie_height / movie_width * 100
+            height_minimap = int(movie_height / movie_width * 100)
             # draw in the upper right corner, overview rectangle
             x = self.width() - length_minimap - 20
             y = 20
@@ -6818,10 +6819,10 @@ class View(QtWidgets.QLabel):
             painter.setPen(QtGui.QColor("yellow"))
             if self.window.dataset_dialog.wbackground.isChecked():
                 painter.setPen(QtGui.QColor("red"))
-            length = self.viewport_width() / movie_width * length_minimap
-            height = self.viewport_height() / movie_height * height_minimap
-            x_vp = self.viewport[0][1] / movie_width * length_minimap
-            y_vp = self.viewport[0][0] / movie_height * length_minimap
+            length = int(self.viewport_width() / movie_width * length_minimap)
+            height = int(self.viewport_height() / movie_height * height_minimap)
+            x_vp = int(self.viewport[0][1] / movie_width * length_minimap)
+            y_vp = int(self.viewport[0][0] / movie_height * length_minimap)
             painter.drawRect(x + x_vp, y + y_vp, length + 0, height + 0)
         return image
 
@@ -7308,7 +7309,7 @@ class View(QtWidgets.QLabel):
 
         cx = self.width() * (x - self.viewport[0][1]) / self.viewport_width()
         cy = self.height() * (y - self.viewport[0][0]) / self.viewport_height()
-        return cx, cy
+        return int(cx), int(cy)
 
     def max_movie_height(self):
         """ Returns maximum height of all loaded images. """
@@ -7733,7 +7734,7 @@ class View(QtWidgets.QLabel):
                     fig, (ax1, ax2, ax3) = plt.subplots(
                         3, 1, figsize=(5, 5), constrained_layout=True
                     )
-                    fig.canvas.set_window_title("Trace")
+                    fig.canvas.manager.set_window_title("Trace")
                     pick = self._picks[i]
                     locs = all_picked_locs[i]
                     locs = stack_arrays(locs, asrecarray=True, usemask=False)
@@ -7844,7 +7845,6 @@ class View(QtWidgets.QLabel):
             raise NotImplementedError(
                 "Not implemented for rectangular picks"
             )
-        print("Showing picks...")
         channel = self.get_channel3d("Select Channel")
 
         removelist = [] # picks to be removed
@@ -7864,7 +7864,9 @@ class View(QtWidgets.QLabel):
                     i = 0
                     while i < len(self._picks):
                         fig = plt.figure(figsize=(5, 5))
-                        fig.canvas.set_window_title("Scatterplot of Pick")
+                        fig.canvas.manager.set_window_title(
+                            "Scatterplot of Pick"
+                        )
                         pick = self._picks[i]
 
                         # plot scatter
@@ -7948,7 +7950,9 @@ class View(QtWidgets.QLabel):
                     while i < len(self._picks):
                         pick = self._picks[i]
                         fig = plt.figure(figsize=(5, 5))
-                        fig.canvas.set_window_title("Scatterplot of Pick")
+                        fig.canvas.manager.set_window_title(
+                            "Scatterplot of Pick"
+                        )
                         ax = fig.add_subplot(111)
                         ax.set_title(
                             "Scatterplot of Pick "
@@ -8031,7 +8035,6 @@ class View(QtWidgets.QLabel):
         Uses PlotDialog for displaying the scatter.
         """
 
-        print("Show pick 3D")
         channel = self.get_channel3d("Show Pick 3D")
         pixelsize = self.window.display_settings_dlg.pixelsize.value()
         removelist = []
@@ -8149,7 +8152,6 @@ class View(QtWidgets.QLabel):
     def analyze_cluster(self):
         """ Clusters picked locs using k-means clustering. """
 
-        print("Analyzing clusters...")
         channel = self.get_channel3d("Show Pick 3D")
         removelist = []
         saved_locs = []
@@ -8356,7 +8358,7 @@ class View(QtWidgets.QLabel):
 
                 # plot histogram with n_locs in picks
                 fig = plt.figure()
-                fig.canvas.set_window_title("Localizations in Picks")
+                fig.canvas.manager.set_window_title("Localizations in Picks")
                 ax = fig.add_subplot(111)
                 ax.set_title("Localizations in Picks ")
                 n, bins, patches = ax.hist(
@@ -10070,7 +10072,7 @@ class View(QtWidgets.QLabel):
                 diameter = (
                     self.window.tools_settings_dialog.pick_diameter.value()
                 )
-                diameter = self.width() * diameter / self.viewport_width()
+                diameter = int(self.width() * diameter / self.viewport_width())
                 # remote desktop crashes sometimes for high diameter
                 if diameter < 100:
                     pixmap_size = ceil(diameter)
@@ -10080,7 +10082,7 @@ class View(QtWidgets.QLabel):
                     painter.setPen(QtGui.QColor("white"))
                     if self.window.dataset_dialog.wbackground.isChecked():
                         painter.setPen(QtGui.QColor("black"))
-                    offset = (pixmap_size - diameter) / 2
+                    offset = int((pixmap_size - diameter) / 2)
                     painter.drawEllipse(offset, offset, diameter, diameter)
                     painter.end()
                     cursor = QtGui.QCursor(pixmap)
@@ -11233,8 +11235,6 @@ class Window(QtWidgets.QMainWindow):
                 self.export_3d_visp()
             elif item == ".csv for ThunderSTORM":
                 self.export_ts()
-            else:
-                print("This should never happen")
 
     def export_ts(self):
         """ Exports locs as .csv for ThunderSTORM. """
@@ -11725,7 +11725,6 @@ class Window(QtWidgets.QMainWindow):
             if input[0] == "flip" and len(input) == 3:
                 # Distinguish flipping in xy and z
                 if "z" in input:
-                    print("xyz")
                     var_1 = input[1]
                     var_2 = input[2]
                     if var_1 == "z":
@@ -11893,7 +11892,6 @@ class Window(QtWidgets.QMainWindow):
         channel = self.view.get_channel_all_seq("Save localizations")
         if channel is not None:
             if channel == len(self.view.locs_paths):
-                print("Save all at once.")
                 suffix, ok = QtWidgets.QInputDialog.getText(
                     self,
                     "Input Dialog",